Update Mde/MdeModulePkg to support ICC build for IA32/X64.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@3892 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/Network/Tcp4Dxe/Tcp4Option.c
@@ -249,7 +249,7 @@ TcpParseOption (
 
   Option->Flag  = 0;
 
-  TotalLen      = (Tcp->HeadLen << 2) - sizeof (TCP_HEAD);
+  TotalLen      = (UINT8) ((Tcp->HeadLen << 2) - sizeof (TCP_HEAD));
   if (TotalLen <= 0) {
     return 0;
   }
